refactor and changes
