+++ config/base/crds/istio_v1beta1_remoteistio.yaml
@@ -17,7 +17,7 @@ spec:
     type: string
   - JSONPath: .status.GatewayAddress
     description: Ingress gateways of the resource
-    name: Gateways
+    name: Ingress IPs
     type: string
   - JSONPath: .metadata.creationTimestamp
     name: Age
@@ -409,6 +409,15 @@ spec:
           - enabledServices
           type: object
         status:
+          properties:
+            ErrorMessage:
+              type: string
+            GatewayAddress:
+              items:
+                type: string
+              type: array
+            Status:
+              type: string
           type: object
   version: v1beta1
 status:
