+++ pkg/apis/istio/v1beta1/istio_types.go
@@ -701,9 +701,9 @@ func (v IstioVersion) IsSupported() bool {
 
 // IstioStatus defines the observed state of Istio
 type IstioStatus struct {
-	Status         ConfigState
-	GatewayAddress []string
-	ErrorMessage   string
+	Status         ConfigState `json:"Status,omitempty"`
+	GatewayAddress []string    `json:"GatewayAddress,omitempty"`
+	ErrorMessage   string      `json:"ErrorMessage,omitempty"`
 }
 
 // +genclient
@@ -714,7 +714,7 @@ type IstioStatus struct {
 // +kubebuilder:subresource:status
 // +kubebuilder:printcolumn:name="Status",type="string",JSONPath=".status.Status",description="Status of the resource"
 // +kubebuilder:printcolumn:name="Error",type="string",JSONPath=".status.ErrorMessage",description="Error message"
-// +kubebuilder:printcolumn:name="Gateways",type="string",JSONPath=".status.GatewayAddress",description="Ingress gateways of the resource"
+// +kubebuilder:printcolumn:name="Ingress IPs",type="string",JSONPath=".status.GatewayAddress",description="Ingress gateways of the resource"
 // +kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp"
 type Istio struct {
 	metav1.TypeMeta   `json:",inline"`
