+++ pkg/apis/istio/v1beta1/remoteistio_types.go
@@ -78,9 +78,9 @@ type RemoteIstioSpec struct {
 
 // RemoteIstioStatus defines the observed state of RemoteIstio
 type RemoteIstioStatus struct {
-	Status         ConfigState
-	GatewayAddress []string
-	ErrorMessage   string
+	Status         ConfigState `json:"Status,omitempty"`
+	GatewayAddress []string    `json:"GatewayAddress,omitempty"`
+	ErrorMessage   string      `json:"ErrorMessage,omitempty"`
 }
 
 // +genclient
@@ -91,7 +91,7 @@ type RemoteIstioStatus struct {
 // +kubebuilder:subresource:status
 // +kubebuilder:printcolumn:name="Status",type="string",JSONPath=".status.Status",description="Status of the resource"
 // +kubebuilder:printcolumn:name="Error",type="string",JSONPath=".status.ErrorMessage",description="Error message"
-// +kubebuilder:printcolumn:name="Gateways",type="string",JSONPath=".status.GatewayAddress",description="Ingress gateways of the resource"
+// +kubebuilder:printcolumn:name="Ingress IPs",type="string",JSONPath=".status.GatewayAddress",description="Ingress gateways of the resource"
 // +kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp"
 type RemoteIstio struct {
 	metav1.TypeMeta   `json:",inline"`
