+++ pkg/remoteclusters/gateway_address.go
@@ -17,13 +17,10 @@ limitations under the License.
 package remoteclusters
 
 import (
-	"context"
-
-	"github.com/pkg/errors"
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	istiov1beta1 "github.com/banzaicloud/istio-operator/pkg/apis/istio/v1beta1"
+	"github.com/banzaicloud/istio-operator/pkg/k8sutil"
 	"github.com/banzaicloud/istio-operator/pkg/resources/ingressgateway"
 	"github.com/banzaicloud/istio-operator/pkg/util"
 )
@@ -33,26 +30,11 @@ func (c *Cluster) SetIngressGatewayAddress(remoteIstio *istiov1beta1.RemoteIstio
 		return nil
 	}
 
-	c.log.Info("get ingress gateway address")
-
-	var mgw istiov1beta1.MeshGateway
-	var ips []string
-
-	err := c.ctrlRuntimeClient.Get(context.Background(), types.NamespacedName{
+	var err error
+	remoteIstio.Status.GatewayAddress, err = k8sutil.GetMeshGatewayAddress(c.ctrlRuntimeClient, client.ObjectKey{
 		Name:      ingressgateway.ResourceName,
 		Namespace: remoteIstio.Namespace,
-	}, &mgw)
-	if err != nil && !k8serrors.IsNotFound(err) {
-		return err
-	}
+	})
 
-	if mgw.Status.Status != istiov1beta1.Available {
-		return errors.New("gateway is pending")
-	}
-
-	ips = mgw.Status.GatewayAddress
-
-	remoteIstio.Status.GatewayAddress = ips
-
-	return nil
+	return err
 }
